aiService/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.9
azure-core==1.32.0
azure-eventhub==5.14.0
azure-identity==1.21.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dill==0.3.9
fastapi==0.115.11
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isort==6.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pylint==3.3.5
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tomlkit==0.13.2
torch==2.6.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
